organized the canvas folder
