+++ src/canvas/Geometry/BestCurveSelector.tsx
