+++ src/canvas/Geometry/BoundingBox.tsx
