+++ src/canvas/Geometry/Curve.tsx
