+++ src/canvas/Geometry/CurveFitter.tsx
