+++ src/canvas/Geometry/Line.tsx
@@ -1,0 +1,3 @@
+export class Segment {
+    
+}
