+++ src/canvas/Geometry/Point.tsx
