+++ src/canvas/Geometry/Segment.tsx
@@ -1,0 +1,3 @@
+export abstract class Segment {
+    
+}
