+++ src/canvas/PatternPaths/StraightLinePath.tsx
